+++ TombEngine/Game/Animation/Animation.cpp
@@ -169,7 +169,6 @@ namespace TEN::Animation
 	{
 		return (Offset == Vector3::Zero &&
 				Rotation == EulerAngles::Identity &&
-				ExtraRotation == EulerAngles::Identity &&
 				Scale == Vector3::One);
 	};